Refactor folder structure and added views to page layout
