+++ eslint.config.mjs
@@ -18,6 +18,7 @@ export default [
   {
     rules: {
       'no-unused-vars': 'warn',
+      'react/prop-types': 'off',
     },
   },
   {
