--- src/views/components/SinglePost.jsx
@@ -1,8 +1,0 @@
-function SinglePost() {
-  return (
-    // <div className="flex justify-between border-b-2 border-neutral-700 bg-neutral-900 px-10 py-3">
-    <div>SinglePost</div>
-  );
-}
-
-export default SinglePost;
