+++ src/views/public-blog/components/Comments.jsx
@@ -1,0 +1,9 @@
+function Comments() {
+  return (
+    <div className="mt-10 flex flex-col justify-start text-black">
+      Comments
+    </div>
+  );
+}
+
+export default Comments;
