+++ src/views/public-blog/components/Legenda.jsx
