+++ src/views/public-blog/pages/CreatePost.jsx
@@ -1,0 +1,9 @@
+function CreatePost() {
+  return (
+    <div className="mt-10 flex flex-col justify-start text-black">
+      CreatePost
+    </div>
+  );
+}
+
+export default CreatePost;
