+++ src/views/public-blog/pages/Register.jsx
@@ -1,0 +1,9 @@
+function Register() {
+  return (
+    <div className="mt-10 flex flex-col justify-start text-black">
+      Register
+    </div>
+  );
+}
+
+export default Register;
