+++ src/views/public-blog/pages/Signin.jsx
@@ -1,0 +1,9 @@
+function Signin() {
+  return (
+    <div className="mt-10 flex flex-col justify-start text-black">
+      Signin
+    </div>
+  );
+}
+
+export default Signin;
